Add SL_MODE macro for upper compatibility with Pgpool-II 3.7 or later.
SL_MODE is streaming replication mode or logical replication mode.
Since in 3.7 or later, SL_MODE is used in all over the place instead
of STREAM. This makes back patching harder. So in 3.6 or earlier, we
define SL_MODE as STREAM so that we can use the patch created in 3.7
or later.