Enhance extended query test.
Add extra_scripts directory to include extra scripts to be executed
after main test script (tests) run.  Currently only scripts for
parse-before-bind.data parse-before-bind-2.data exist so that the test
confirm the load balancing behavior.

Also some tests are fixed so that they can run individually in that a
table for testing is dropped at the end of the test and test results
are not affected by the existence of the table.