Allow to display Pgpool-II child process id and PostgreSQL backend id in pcp_proc_info.
If --all option was not specified, Pgpool-II child process id was not
displayed. This is not reasonable since if Pgpool-II child process id
is not given, multiple entries of Pgpool-II child process are
displayed without no information to distinguish them.

So fix the command so that it always display Pgpool-II child process
id. Also there's no reason not to display PostgreSQL backend id, so
displays it.