Add new log_client_messages directive.
This allows to log messages from frontend.

Note that the regression test for this (019.log_client_messages)
requires pgproto command.  We will eventually import pgproto in the
future but for now we just check the existence of pgproto in
019.log_client_messages/test.sh. If pgproto does not exist in the
command search pass, just skip the test with "ok" status.

Patch contributed by Takuma Hoshiai.