Downgrade warning to debug5 when AES/PLAIN password not found in pool_passwd.
When health_check_password etc. is an empty string, Pgpool-II tries to
fetch AES/PLAIN password in pool_passwd. If other type of password
entry (actually it must be MD5), it warned it. However this is normal
setting till 3.7. Moreover, every time health check or streaming
replication delay check run, it emits the waring and the log file was
filled up with the message.

Solution is, downgrade the message to DEBUG5. Also add password type to
the message.