Try to avoid 069.memory_leak_extended test on 3.5 branch.
It is strange that it only happens on 3.5 branch and in raw mode.  One
theory is, ps does not quickly shows process size after running
pgbench -S -T 1 to get the initial process size. So tweak 1 to 10 to
give ps a chance to update the status. See if this works.