Fix for 0000419: pg_md5: -m removes users
The problem was in pool_create_passwdent() which was expecting the password
string will always have the same length. This is not true anymore after the
support for AES and TEXT type passwords, So inlace updating of pool_passwd is
not an option anymore.
The fix is to use the memory buffer to compile the pool_passwd file and rewrite
it completely.
The patch was provided by Takuma Hoshiai <hoshiai@sraoss.co.jp>
and modified by me.