Enhance performance while sending message to frontend.
SimpleForwardToFrontend(), which is responsible for sending message to
frontend, does write buffering only if it is either 'D' (DataRow) or
'd' (CopyData). Other message types were immediately written to
socket. But actually this was not necessary. So if the messages are
not critical ("Command Complete", "Ready For query", + * "Error
response" and "Notice message" messages), just write to buffer.  With
this 10-17% performance enhance was observed.

Discussion: [pgpool-hackers: 3233] https://www.pgpool.net/pipermail/pgpool-hackers/2019-February/003233.html