Add new 'enable_shared_relcache' parameter.
The relation cache were stored in local cache of child processes, so all child processes executed same query to get relation cache.
If enable_shared_relcache is on, the relation cache is stored in memory cache and all child process share it.
It will expect to reduce the load that same query is executed.