Set application_name in recovery_conf.
This allows to show each cluster name as application_name. Currently
each application_name is set to: server0, server1, server2 and so on,
where the numeric part is the cluster number. This is especially
useful to use pg_stat_replication because it does not provide any useful
information regarding UNIX socket path or port number.