Allow to route relcache queries to load balance node.
Queries to build relcache entries were always sent to master (primary)
node. This is usually good because we could eliminate the bad effect
of replication delay. However if users want to lower the load of
master node, it would be nice if we could route the queries to other
than master node. This patch introduces new parameter
"relache_query_target". If it is set to 'load_balance_node', relcache
queries will be routed to load balance node.  If it is set to
'master', the queries are routed to master node, which is same as
before (this is the default).