Feature: auto failback
In Pgpool-II 4.1 or later, you can reattach backend nodethat is status DOWN automatically if auto_failback set on, this future is enabled.

The Feature of auto_failback require that backend node of PostgreSQL are 9.1 or later with streaming-replication mode, and sr_check process and health_check is enabled.

It do health check for standby node, when status of standby Node is down but status of streaming replication between primary and standby is 'stream'. If health check is success, execute process of fail_back automatically. This featurel is usefull for health check is failed bacause of tempolary network error is happened for example.